@@ -3,7 +3,7 @@ require "rails/generators"
 module Pghero
   module Generators
     class ConfigGenerator < Rails::Generators::Base
-      source_root File.expand_path("../templates", __FILE__)
+      source_root File.join(__dir__, "templates")
 
       def create_initializer
         template "config.yml", "config/pghero.yml"
