@@ -1,34 +1,17 @@
-# taken from https://github.com/collectiveidea/audited/blob/master/lib/generators/audited/install_generator.rb
-require "rails/generators"
-require "rails/generators/migration"
-require "active_record"
 require "rails/generators/active_record"
 
 module Pghero
   module Generators
     class SpaceStatsGenerator < Rails::Generators::Base
-      include Rails::Generators::Migration
-
-      source_root File.expand_path("../templates", __FILE__)
-
-      # Implement the required interface for Rails::Generators::Migration.
-      def self.next_migration_number(dirname) #:nodoc:
-        next_migration_number = current_migration_number(dirname) + 1
-        if ::ActiveRecord::Base.timestamped_migrations
-          [Time.now.utc.strftime("%Y%m%d%H%M%S"), "%.14d" % next_migration_number].max
-        else
-          "%.3d" % next_migration_number
-        end
-      end
+      include ActiveRecord::Generators::Migration
+      source_root File.join(__dir__, "templates")
 
       def copy_migration
         migration_template "space_stats.rb", "db/migrate/create_pghero_space_stats.rb", migration_version: migration_version
       end
 
       def migration_version
-        if ActiveRecord::VERSION::MAJOR >= 5
-          "[#{ActiveRecord::VERSION::MAJOR}.#{ActiveRecord::VERSION::MINOR}]"
-        end
+        "[#{ActiveRecord::VERSION::MAJOR}.#{ActiveRecord::VERSION::MINOR}]"
       end
     end
   end
