@@ -5,8 +5,8 @@ module PgHero
         select_one("SELECT pg_terminate_backend(#{pid.to_i})")
       end
 
-      def kill_long_running_queries
-        long_running_queries.each { |query| kill(query[:pid]) }
+      def kill_long_running_queries(min_duration: nil)
+        running_queries(min_duration: min_duration || long_running_query_sec).each { |query| kill(query[:pid]) }
         true
       end
 
