@@ -5,6 +5,8 @@ module PgHero
         # get columns with default values
         # use pg_get_expr to get correct default value
         # it's what information_schema.columns uses
+        # also, exclude temporary tables to prevent error
+        # when accessing across sessions
         sequences = select_all <<-SQL
           SELECT
             n.nspname AS table_schema,
@@ -23,7 +25,8 @@ module PgHero
           WHERE
             NOT a.attisdropped
             AND a.attnum > 0
-            AND d.adsrc LIKE 'nextval%'
+            AND pg_get_expr(d.adbin, d.adrelid) LIKE 'nextval%'
+            AND n.nspname NOT LIKE 'pg\\_temp\\_%'
         SQL
 
         # parse out sequence
