@@ -1,3 +1,3 @@
 module PgHero
-  VERSION = "2.3.0"
+  VERSION = "2.4.1"
 end
